remove paddle_serving_server_gpu
